pyproject.toml
[tool.poetry]
name = "anki-tutor"
version = "0.1.0"
description = ""
authors = ["Adam Szustak <as.szustak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
speechrecognition = "^3.10.4"
pyaudio = "^0.2.14"
setuptools = "^70.0.0"
torch = "2.2.0"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
openai = "^1.31.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.10.0"
black = "^24.4.2"
isort = "^5.13.2"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_equality = true
check_untyped_defs = true
disallow_untyped_decorators = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
warn_return_any = true

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
)/
'''

[tool.isort]
profile = "black"
atomic = true
line_length = 79
skip_gitignore = true
skip_glob = ["mypy_cache"]
known_first_party = ["anki-tutor", "llm", "db", "exceptions"]
include_trailing_comma = true
multi_line_output = 3
force_grid_wrap = 2
